--- scenario_player/exceptions.py
@@ -1,34 +1,0 @@
-class ScenarioError(Exception):
-    pass
-
-
-class ScenarioTxError(ScenarioError):
-    pass
-
-
-class TokenRegistrationError(ScenarioTxError):
-    pass
-
-
-class TransferFailed(ScenarioError):
-    pass
-
-
-class RESTAPIError(ScenarioError):
-    pass
-
-
-class RESTAPIStatusMismatchError(ScenarioError):
-    pass
-
-
-class RESTAPITimeout(RESTAPIError):
-    pass
-
-
-class UnknownTaskTypeError(ScenarioError):
-    pass
-
-
-class ScenarioAssertionError(ScenarioError):
-    pass
